+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # Change Log
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
-## [1-alpha2-SNAPSHOT]
+## [1-alpha2] - 2022-02-21
 - a more graceful return value
 - added CSS to dataset view
+- added :kind/table support using Clerk's table view
 
 ## [1-alpha1] - 2022-02-03
 - initial working version